backend/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
django-filter==24.3
django-environ==0.11.2
djangorestframework-simplejwt==5.3.1
stripe==10.2.0
sendgrid==6.11.0
et_xmlfile==2.0.0
gunicorn==23.0.0
openpyxl==3.1.5
packaging==25.0
prometheus-client==0.20.0
sqlparse==0.5.3
tzdata==2025.2
waitress==3.0.2
pytest==8.3.3
pytest-django==4.9.0
factory-boy==3.3.0
psycopg[binary]>=3.2
openai==1.54.1
jsonschema==4.23.0
